Add dashboard loading skeletons
